fix: merge authentisation error to error class files
