--- backend/src/errors/AuthenticationError.ts
@@ -1,9 +1,0 @@
-export class AuthenticationError extends Error {
-  status: number;
-
-  constructor(message: string) {
-    super(message);
-    this.name = 'AuthenticationError';
-    this.status = 401;
-  }
-}
